MDCStudio: ImageView: rendering full-size image working

+++ Tools/MDCStudio/Source/ImagePipeline/ImagePipeline.h
@@ -12,8 +12,8 @@ class Pipeline {
 public:
     struct RawImage {
         CFADesc cfaDesc;
-        uint32_t width = 0;
-        uint32_t height = 0;
+        size_t width = 0;
+        size_t height = 0;
         const ImagePixel* pixels = nullptr;
     };
